BOLT: Split out game code into their own folders
