+++ engines/bolt/bolt.cpp
@@ -39,8 +39,6 @@ BoltEngine::BoltEngine(OSystem *syst, const ADGameDescription *gameDesc) : Engin
 	_gameDescription(gameDesc), _randomSource("Bolt") {
 	g_engine = this;
 	_xp = new XpLib(this);
-
-	initCallbacks();
 }
 
 BoltEngine::~BoltEngine() {
@@ -81,120 +79,6 @@ Common::Error BoltEngine::run() {
 	return Common::kNoError;
 }
 
-void BoltEngine::boltMain() {
-	byte *testAlloc;
-	BarkerTable *barkerTable;
-	byte *boothSprite;
-
-	testAlloc = (byte *)_xp->allocMem(0x100000);
-	if (!testAlloc)
-		return;
-
-	_xp->freeMem(testAlloc);
-
-	_xp->randomize();
-
-	if (allocResourceIndex()) {
-		_boothsBoltLib = nullptr;
-
-		if (openBOLTLib(&_boothsBoltLib, &_boothsBoltCallbacks, assetPath("booths.blt"))) {
-			int16 chosenSpecId = _extendedViewport ? 0 : 1;
-
-			if (_xp->setDisplaySpec(&_displayMode, &_displaySpecs[chosenSpecId])) {
-				assert(_displayMode >= 0);
-				_displayWidth = _displaySpecs[_displayMode].width;
-				_displayHeight = _displaySpecs[_displayMode].height;
-
-				// Center within 384x240 virtual coordinate space...
-				_displayX = (EXTENDED_SCREEN_WIDTH - _displayWidth) / 2;
-				_displayY = (EXTENDED_SCREEN_HEIGHT - _displayHeight) / 2;
-
-				_xp->setCoordSpec(_displayX, _displayY, _displayWidth, _displayHeight);
-				
-				if (_displayMode == 0) {
-					_rtfHandle = openRTF(assetPath("cc_og.av"));
-				} else {
-					_rtfHandle = openRTF(assetPath("cc_cr.av"));
-				}
-
-				if (_rtfHandle) {
-					int16 boothGroupBase = _isDemo ? 0x0E00 : 0x1700;
-
-					if (!_isDemo) {
-						playAV(_rtfHandle, 0, _displayWidth, _displayHeight, _displayX, _displayY);
-					}
-
-					boothSprite = getBOLTMember(_boothsBoltLib, (_displayMode != 0) ? (boothGroupBase + 1) : (boothGroupBase + 2));
-
-					_xp->setTransparency(false);
-
-					displayColors(getBOLTMember(_boothsBoltLib, boothGroupBase), stFront, 0);
-					displayPic(boothSprite, _displayX, _displayY, stFront);
-
-					_xp->updateDisplay();
-
-					displayColors(getBOLTMember(_boothsBoltLib, boothGroupBase), stBack, 0);
-					displayPic(boothSprite, _displayX, _displayY, stBack);
-
-					playAV(_rtfHandle, _isDemo ? 0 : 2, _displayWidth, _displayHeight, _displayX, _displayY);
-
-					freeBOLTGroup(_boothsBoltLib, boothGroupBase, 1);
-
-					if (getBOLTGroup(_boothsBoltLib, 0, 1)) {
-						setCursorPict(memberAddr(_boothsBoltLib, 0));
-
-						if (initVRam(1500)) {
-							barkerTable = createBarker(2, _isDemo ? 19 : 17);
-							if (barkerTable) {
-								// Register booth handlers...
-								registerSideShow(barkerTable, &BoltEngine::hucksBooth, 3);
-								registerSideShow(barkerTable, &BoltEngine::fredsBooth, 4);
-								registerSideShow(barkerTable, &BoltEngine::scoobysBooth, 5);
-								registerSideShow(barkerTable, &BoltEngine::yogisBooth, 6);
-								registerSideShow(barkerTable, &BoltEngine::georgesBooth, 7);
-								registerSideShow(barkerTable, &BoltEngine::topCatsBooth, 8);
-								registerSideShow(barkerTable, &BoltEngine::mainEntrance, 9);
-								registerSideShow(barkerTable, &BoltEngine::huckGame, 10);
-								registerSideShow(barkerTable, &BoltEngine::fredGame, 11);
-								registerSideShow(barkerTable, &BoltEngine::scoobyGame, 12);
-								registerSideShow(barkerTable, &BoltEngine::yogiGame, 13);
-								registerSideShow(barkerTable, &BoltEngine::georgeGame, 14);
-								registerSideShow(barkerTable, &BoltEngine::topCatGame, 15);
-								registerSideShow(barkerTable, &BoltEngine::winALetter, 16);
-
-								if (_isDemo) {
-									registerSideShow(barkerTable, &BoltEngine::displayDemoPict, 17);
-									registerSideShow(barkerTable, &BoltEngine::endDemo, 18);
-								}
-
-								_lettersWon = 0;
-								_xp->setScreenSaverTimer(1800);
-
-								// The barker function runs the main loop, starting at mainEntrance()...
-								if (!checkError())
-									barker(barkerTable, 9);
-
-								freeBarker(barkerTable);
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-
-	freeBOLTGroup(_boothsBoltLib, 0, 1);
-
-	if (_boothsBoltLib != 0)
-		closeBOLTLib(&_boothsBoltLib);
-
-	if (_rtfHandle != nullptr)
-		closeRTF(_rtfHandle);
-
-	freeVRam();
-	freeResourceIndex();
-}
-
 void BoltEngine::setCursorPict(byte *sprite) {
 	byte cursorBitmap[32]; // 16x16
 	byte *srcPtr = getResolvedPtr(sprite, 0x12);
@@ -212,27 +96,6 @@ void BoltEngine::setCursorPict(byte *sprite) {
 	}
 
 	_xp->setCursorImage(cursorBitmap, 7, 7);
-}
-
-int16 BoltEngine::displayDemoPict(int16 prevBooth) {
-	uint32 eventData;
-
-	while (_xp->getEvent(etMouseDown, &eventData) == etMouseDown);
-
-	while (_xp->getEvent(etTimer, &eventData) != etTimer);
-
-	_xp->stopCycle();
-	_xp->setTransparency(false);
-
-	int16 memberId = (_displayMode != 0) ? 0x1002 : 0x1001;
-	displayPic(getBOLTMember(_boothsBoltLib, memberId), _displayX, _displayY, stFront);
-	displayColors(getBOLTMember(_boothsBoltLib, 0x1000), stFront, 0);
-
-	_xp->updateDisplay();
-
-	while (_xp->getEvent(etMouseDown, &eventData) != etMouseDown);
-
-	return prevBooth;
 }
 
 } // End of namespace Bolt
