+++ engines/bolt/resource.cpp
@@ -640,24 +640,6 @@ BOLTCallback BoltEngine::_defaultMemberFreeCallbacks[25];
 BOLTCallback BoltEngine::_defaultGroupLoadCallbacks[25];
 BOLTCallback BoltEngine::_defaultGroupFreeCallbacks[25];
 
-BOLTCallback BoltEngine::_fredTypeLoadCallbacks[28];
-BOLTCallback BoltEngine::_fredTypeFreeCallbacks[28];
-
-BOLTCallback BoltEngine::_georgeTypeLoadCallbacks[28];
-BOLTCallback BoltEngine::_georgeTypeFreeCallbacks[28];
-
-BOLTCallback BoltEngine::_huckTypeLoadCallbacks[27];
-BOLTCallback BoltEngine::_huckTypeFreeCallbacks[27];
-
-BOLTCallback BoltEngine::_scoobyTypeLoadCallbacks[28];
-BOLTCallback BoltEngine::_scoobyTypeFreeCallbacks[28];
-
-BOLTCallback BoltEngine::_topCatTypeLoadCallbacks[26];
-BOLTCallback BoltEngine::_topCatTypeFreeCallbacks[26];
-
-BOLTCallback BoltEngine::_yogiTypeLoadCallbacks[27];
-BOLTCallback BoltEngine::_yogiTypeFreeCallbacks[27];
-
 void BoltEngine::noOpCb() {}
 void BoltEngine::swapAllWordsCb() { ((BoltEngine *)g_engine)->swapAllWords(); }
 void BoltEngine::swapAllLongsCb() { ((BoltEngine *)g_engine)->swapAllLongs(); }
@@ -668,231 +650,5 @@ void BoltEngine::swapFirstTwoWordsCb() { ((BoltEngine *)g_engine)->swapFirstTwoW
 void BoltEngine::swapFirstFourWordsCb() { ((BoltEngine *)g_engine)->swapFirstFourWords(); }
 void BoltEngine::swapSpriteHeaderCb() { ((BoltEngine *)g_engine)->swapSpriteHeader(); }
 void BoltEngine::freeSpriteCleanUpCb() { ((BoltEngine *)g_engine)->freeSpriteCleanUp(); }
-
-void BoltEngine::resolveAllRefsCb() { ((BoltEngine *)g_engine)->resolveAllRefs(); }
-void BoltEngine::swapFredAnimEntryCb() { ((BoltEngine *)g_engine)->swapFredAnimEntry(); }
-void BoltEngine::swapFredAnimDescCb() { ((BoltEngine *)g_engine)->swapFredAnimDesc(); }
-void BoltEngine::swapFredLevelDescCb() { ((BoltEngine *)g_engine)->swapFredLevelDesc(); }
-
-void BoltEngine::swapGeorgeFrameArrayCb() { ((BoltEngine *)g_engine)->swapGeorgeFrameArray(); }
-void BoltEngine::swapGeorgeHelpEntryCb() { ((BoltEngine *)g_engine)->swapGeorgeHelpEntry(); }
-void BoltEngine::swapGeorgeThresholdsCb() { ((BoltEngine *)g_engine)->swapGeorgeThresholds(); }
-
-void BoltEngine::swapHuckWordArrayCb() { ((BoltEngine *)g_engine)->swapHuckWordArray(); }
-void BoltEngine::swapHuckWordsCb() { ((BoltEngine *)g_engine)->swapHuckWords(); }
-
-void BoltEngine::swapScoobyHelpEntryCb() { ((BoltEngine *)g_engine)->swapScoobyHelpEntry(); }
-void BoltEngine::swapScoobyWordArrayCb() { ((BoltEngine *)g_engine)->swapScoobyWordArray(); }
-
-void BoltEngine::swapTopCatHelpEntryCb() { ((BoltEngine *)g_engine)->swapTopCatHelpEntry(); }
-
-void BoltEngine::swapYogiAllWordsCb() { ((BoltEngine *)g_engine)->swapYogiAllWords(); }
-void BoltEngine::swapYogiFirstWordCb() { ((BoltEngine *)g_engine)->swapYogiFirstWord(); }
-
-void BoltEngine::initCallbacks() {
-	// --- BOOTHS ---
-	for (int i = 0; i < ARRAYSIZE(_defaultTypeLoadCallbacks); i++) {
-		_defaultTypeLoadCallbacks[i] = noOpCb;
-	}
-
-	_defaultTypeLoadCallbacks[2] = swapAllWordsCb;
-	_defaultTypeLoadCallbacks[8] = swapSpriteHeaderCb;
-	_defaultTypeLoadCallbacks[10] = swapPicHeaderCb;
-	_defaultTypeLoadCallbacks[11] = swapAndResolvePicDescCb;
-	_defaultTypeLoadCallbacks[12] = swapFirstTwoWordsCb;
-	_defaultTypeLoadCallbacks[14] = swapFirstFourWordsCb;
-
-	for (int i = 0; i < ARRAYSIZE(_defaultTypeFreeCallbacks); i++) {
-		_defaultTypeFreeCallbacks[i] = noOpCb;
-	}
-
-	_defaultTypeFreeCallbacks[8] = freeSpriteCleanUpCb;
-
-	for (int i = 0; i < ARRAYSIZE(_defaultMemberLoadCallbacks); i++) {
-		_defaultMemberLoadCallbacks[i] = noOpCb;
-	}
-
-	for (int i = 0; i < ARRAYSIZE(_defaultMemberFreeCallbacks); i++) {
-		_defaultMemberFreeCallbacks[i] = noOpCb;
-	}
-
-	for (int i = 0; i < ARRAYSIZE(_defaultGroupLoadCallbacks); i++) {
-		_defaultGroupLoadCallbacks[i] = noOpCb;
-	}
-
-	for (int i = 0; i < ARRAYSIZE(_defaultGroupFreeCallbacks); i++) {
-		_defaultGroupFreeCallbacks[i] = noOpCb;
-	}
-
-	_boothsBoltCallbacks.typeLoadCallbacks = _defaultTypeLoadCallbacks;
-	_boothsBoltCallbacks.typeFreeCallbacks = _defaultTypeFreeCallbacks;
-	_boothsBoltCallbacks.memberLoadCallbacks = _defaultMemberLoadCallbacks;
-	_boothsBoltCallbacks.memberFreeCallbacks = _defaultMemberFreeCallbacks;
-	_boothsBoltCallbacks.groupLoadCallbacks = _defaultGroupLoadCallbacks;
-	_boothsBoltCallbacks.groupFreeCallbacks = _defaultGroupFreeCallbacks;
-
-	// --- FRED ---
-	for (int i = 0; i < ARRAYSIZE(_fredTypeLoadCallbacks); i++) {
-		_fredTypeLoadCallbacks[i] = noOpCb;
-	}
-
-	_fredTypeLoadCallbacks[2] = swapAllWordsCb;
-	_fredTypeLoadCallbacks[6] = resolveAllRefsCb;
-	_fredTypeLoadCallbacks[8] = swapSpriteHeaderCb;
-	_fredTypeLoadCallbacks[10] = swapPicHeaderCb;
-	_fredTypeLoadCallbacks[11] = swapAndResolvePicDescCb;
-	_fredTypeLoadCallbacks[12] = swapFirstTwoWordsCb;
-	_fredTypeLoadCallbacks[14] = swapFirstFourWordsCb;
-	_fredTypeLoadCallbacks[25] = swapFredAnimEntryCb;
-	_fredTypeLoadCallbacks[26] = swapFredAnimDescCb;
-	_fredTypeLoadCallbacks[27] = swapFredLevelDescCb;
-
-	for (int i = 0; i < ARRAYSIZE(_fredTypeFreeCallbacks); i++) {
-		_fredTypeFreeCallbacks[i] = noOpCb;
-	}
-
-	_fredTypeFreeCallbacks[8] = freeSpriteCleanUpCb;
-
-	_fredBoltCallbacks.typeLoadCallbacks = _fredTypeLoadCallbacks;
-	_fredBoltCallbacks.typeFreeCallbacks = _fredTypeFreeCallbacks;
-	_fredBoltCallbacks.memberLoadCallbacks = _defaultMemberLoadCallbacks;
-	_fredBoltCallbacks.memberFreeCallbacks = _defaultMemberFreeCallbacks;
-	_fredBoltCallbacks.groupLoadCallbacks = _defaultGroupLoadCallbacks;
-	_fredBoltCallbacks.groupFreeCallbacks = _defaultGroupFreeCallbacks;
-
-	// --- GEORGE ---
-	for (int i = 0; i < ARRAYSIZE(_georgeTypeLoadCallbacks); i++) {
-		_georgeTypeLoadCallbacks[i] = noOpCb;
-	}
-
-	_georgeTypeLoadCallbacks[2] = swapAllWordsCb;
-	_georgeTypeLoadCallbacks[6] = resolveAllRefsCb;
-	_georgeTypeLoadCallbacks[8] = swapSpriteHeaderCb;
-	_georgeTypeLoadCallbacks[10] = swapPicHeaderCb;
-	_georgeTypeLoadCallbacks[11] = swapAndResolvePicDescCb;
-	_georgeTypeLoadCallbacks[12] = swapFirstTwoWordsCb;
-	_georgeTypeLoadCallbacks[14] = swapFirstFourWordsCb;
-	_georgeTypeLoadCallbacks[25] = swapGeorgeFrameArrayCb;
-	_georgeTypeLoadCallbacks[26] = swapGeorgeHelpEntryCb;
-	_georgeTypeLoadCallbacks[27] = swapGeorgeThresholdsCb;
-
-	for (int i = 0; i < ARRAYSIZE(_georgeTypeFreeCallbacks); i++) {
-		_georgeTypeFreeCallbacks[i] = noOpCb;
-	}
-
-	_georgeTypeFreeCallbacks[8] = freeSpriteCleanUpCb;
-
-	_georgeBoltCallbacks.typeLoadCallbacks = _georgeTypeLoadCallbacks;
-	_georgeBoltCallbacks.typeFreeCallbacks = _georgeTypeFreeCallbacks;
-	_georgeBoltCallbacks.memberLoadCallbacks = _defaultMemberLoadCallbacks;
-	_georgeBoltCallbacks.memberFreeCallbacks = _defaultMemberFreeCallbacks;
-	_georgeBoltCallbacks.groupLoadCallbacks = _defaultGroupLoadCallbacks;
-	_georgeBoltCallbacks.groupFreeCallbacks = _defaultGroupFreeCallbacks;
-
-	// --- HUCK ---
-	for (int i = 0; i < ARRAYSIZE(_huckTypeLoadCallbacks); i++) {
-		_huckTypeLoadCallbacks[i] = noOpCb;
-	}
-
-	_huckTypeLoadCallbacks[2] = swapAllWordsCb;
-	_huckTypeLoadCallbacks[6] = resolveAllRefsCb;
-	_huckTypeLoadCallbacks[8] = swapSpriteHeaderCb;
-	_huckTypeLoadCallbacks[10] = swapPicHeaderCb;
-	_huckTypeLoadCallbacks[11] = swapAndResolvePicDescCb;
-	_huckTypeLoadCallbacks[12] = swapFirstTwoWordsCb;
-	_huckTypeLoadCallbacks[14] = swapFirstFourWordsCb;
-	_huckTypeLoadCallbacks[25] = swapHuckWordArrayCb;
-	_huckTypeLoadCallbacks[26] = swapHuckWordsCb;
-
-	for (int i = 0; i < ARRAYSIZE(_huckTypeFreeCallbacks); i++) {
-		_huckTypeFreeCallbacks[i] = noOpCb;
-	}
-
-	_huckTypeFreeCallbacks[8] = freeSpriteCleanUpCb;
-
-	_huckBoltCallbacks.typeLoadCallbacks = _huckTypeLoadCallbacks;
-	_huckBoltCallbacks.typeFreeCallbacks = _huckTypeFreeCallbacks;
-	_huckBoltCallbacks.memberLoadCallbacks = _defaultMemberLoadCallbacks;
-	_huckBoltCallbacks.memberFreeCallbacks = _defaultMemberFreeCallbacks;
-	_huckBoltCallbacks.groupLoadCallbacks = _defaultGroupLoadCallbacks;
-	_huckBoltCallbacks.groupFreeCallbacks = _defaultGroupFreeCallbacks;
-
-	// --- SCOOBY ---
-	for (int i = 0; i < ARRAYSIZE(_scoobyTypeLoadCallbacks); i++) {
-		_scoobyTypeLoadCallbacks[i] = noOpCb;
-	}
-
-	_scoobyTypeLoadCallbacks[8] = swapSpriteHeaderCb;
-	_scoobyTypeLoadCallbacks[10] = swapPicHeaderCb;
-	_scoobyTypeLoadCallbacks[11] = swapAndResolvePicDescCb;
-	_scoobyTypeLoadCallbacks[12] = swapFirstTwoWordsCb;
-	_scoobyTypeLoadCallbacks[26] = swapScoobyHelpEntryCb;
-	_scoobyTypeLoadCallbacks[27] = swapScoobyWordArrayCb;
-
-	for (int i = 0; i < ARRAYSIZE(_scoobyTypeFreeCallbacks); i++) {
-		_scoobyTypeFreeCallbacks[i] = noOpCb;
-	}
-
-	_scoobyTypeFreeCallbacks[8] = freeSpriteCleanUpCb;
-
-	_scoobyBoltCallbacks.typeLoadCallbacks = _scoobyTypeLoadCallbacks;
-	_scoobyBoltCallbacks.typeFreeCallbacks = _scoobyTypeFreeCallbacks;
-	_scoobyBoltCallbacks.memberLoadCallbacks = _defaultMemberLoadCallbacks;
-	_scoobyBoltCallbacks.memberFreeCallbacks = _defaultMemberFreeCallbacks;
-	_scoobyBoltCallbacks.groupLoadCallbacks = _defaultGroupLoadCallbacks;
-	_scoobyBoltCallbacks.groupFreeCallbacks = _defaultGroupFreeCallbacks;
-
-	// --- TOPCAT ---
-	for (int i = 0; i < ARRAYSIZE(_topCatTypeLoadCallbacks); i++) {
-		_topCatTypeLoadCallbacks[i] = noOpCb;
-	}
-
-	_topCatTypeLoadCallbacks[2] = swapAllWordsCb;
-	_topCatTypeLoadCallbacks[4] = swapAllLongsCb;
-	_topCatTypeLoadCallbacks[6] = resolveAllRefsCb;
-	_topCatTypeLoadCallbacks[8] = swapSpriteHeaderCb;
-	_topCatTypeLoadCallbacks[10] = swapPicHeaderCb;
-	_topCatTypeLoadCallbacks[11] = swapAndResolvePicDescCb;
-	_topCatTypeLoadCallbacks[12] = swapFirstTwoWordsCb;
-	_topCatTypeLoadCallbacks[25] = swapTopCatHelpEntryCb;
-
-	for (int i = 0; i < ARRAYSIZE(_topCatTypeFreeCallbacks); i++) {
-		_topCatTypeFreeCallbacks[i] = noOpCb;
-	}
-
-	_topCatTypeFreeCallbacks[8] = freeSpriteCleanUpCb;
-
-	_topCatBoltCallbacks.typeLoadCallbacks = _topCatTypeLoadCallbacks;
-	_topCatBoltCallbacks.typeFreeCallbacks = _topCatTypeFreeCallbacks;
-	_topCatBoltCallbacks.memberLoadCallbacks = _defaultMemberLoadCallbacks;
-	_topCatBoltCallbacks.memberFreeCallbacks = _defaultMemberFreeCallbacks;
-	_topCatBoltCallbacks.groupLoadCallbacks = _defaultGroupLoadCallbacks;
-	_topCatBoltCallbacks.groupFreeCallbacks = _defaultGroupFreeCallbacks;
-
-	// --- YOGI ---
-	for (int i = 0; i < ARRAYSIZE(_yogiTypeLoadCallbacks); i++) {
-		_yogiTypeLoadCallbacks[i] = noOpCb;
-	}
-
-	_yogiTypeLoadCallbacks[2] = swapAllWordsCb;
-	_yogiTypeLoadCallbacks[8] = swapSpriteHeaderCb;
-	_yogiTypeLoadCallbacks[10] = swapPicHeaderCb;
-	_yogiTypeLoadCallbacks[14] = swapFirstFourWordsCb;
-	_yogiTypeLoadCallbacks[25] = swapYogiFirstWordCb;
-	_yogiTypeLoadCallbacks[26] = swapYogiAllWordsCb;
-
-	for (int i = 0; i < ARRAYSIZE(_yogiTypeFreeCallbacks); i++) {
-		_yogiTypeFreeCallbacks[i] = noOpCb;
-	}
-
-	_yogiTypeFreeCallbacks[8] = freeSpriteCleanUpCb;
-
-	_yogiBoltCallbacks.typeLoadCallbacks = _yogiTypeLoadCallbacks;
-	_yogiBoltCallbacks.typeFreeCallbacks = _yogiTypeFreeCallbacks;
-	_yogiBoltCallbacks.memberLoadCallbacks = _defaultMemberLoadCallbacks;
-	_yogiBoltCallbacks.memberFreeCallbacks = _defaultMemberFreeCallbacks;
-	_yogiBoltCallbacks.groupLoadCallbacks = _defaultGroupLoadCallbacks;
-	_yogiBoltCallbacks.groupFreeCallbacks = _defaultGroupFreeCallbacks;
-}
 
 } // End of namespace Bolt
